QtWebKit doesn't support NPAPI for DirectFB https://bugs.webkit.org/show_bug.cgi?id=36171
Reviewed by NOBODY (OOPS!).

This contribution enables windowless NPAPI support for Qt/WebKit when
using DirectFB. It's been verified to work on MIPS hardware with
Flashlite.

The code is largely based on the Qt/X11 implementation of windowless
NPAPI. It adds the define XP_DFB when enabled to differentiate between
X11 and DirectFB when XP_UNIX is defined. This means that npapi.h will
function the same way for X11 builds.

Since the normal use case for NPAPI on DirectFB is for video on sites
like youtube it's important to make sure the plugin gets an updated
global geometry if the window moves or resizes. This is due to the
fact that the video will be rendered into a different layer rather
than actually into the offscreen surface that's being rendered to.

I have a manual test case that I could help people set up for testing
with Qt/QWS DirectFB but I don't know how to make it automated.

To enable NPAPI for DirectFB the developer needs to make sure
qmake's CONFIG variable contains: directfb-npapi

Qt must also be configured with the -qt-gfx-directfb option (as
opposed to -plugin-gfx-directfb)

Finally qmake's QT_CFLAGS_DIRECTFB and QT_LIBS_DIRECTFB variables must
contain all necessary switches to link against DirectFB.

When building from WebKit trunk this can be achieved doing something
like this:

build-webkit -qt -config directfb-npapi \
    QT_CFLAGS_DIRECTFB=\"`directfb-config --cflags`\" \
    QT_LIBS_DIRECTFB=\"`directfb-config --libs`\"